tams-model/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.104.1
gotrue==1.3.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
joblib==1.3.2
numpy==2.2.6
openpyxl==3.1.2
packaging==25.0
pandas==2.3.1
postgrest==0.13.2
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
realtime==1.0.6
requests==2.31.0
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.2
supafunc==0.3.3
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==12.0
